add dedicated runner to libfuzzer, update docker
